Extract a common class to reuse the same code between S3 and GCS credential vending IT
