+++ iceberg/iceberg-rest-server/src/test/java/org/apache/gravitino/iceberg/integration/test/IcebergRESTS3TokenAuthorizationIT.java
@@ -19,22 +19,9 @@
 
 package org.apache.gravitino.iceberg.integration.test;
 
-import com.google.common.collect.ImmutableList;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.UUID;
-import org.apache.gravitino.MetadataObject;
-import org.apache.gravitino.MetadataObjects;
-import org.apache.gravitino.NameIdentifier;
-import org.apache.gravitino.Namespace;
-import org.apache.gravitino.authorization.Owner;
-import org.apache.gravitino.authorization.Privileges;
-import org.apache.gravitino.authorization.SecurableObject;
-import org.apache.gravitino.authorization.SecurableObjects;
 import org.apache.gravitino.catalog.lakehouse.iceberg.IcebergConstants;
 import org.apache.gravitino.credential.CredentialConstants;
 import org.apache.gravitino.credential.S3TokenCredential;
@@ -42,20 +29,14 @@ import org.apache.gravitino.integration.test.util.BaseIT;
 import org.apache.gravitino.integration.test.util.DownloaderUtils;
 import org.apache.gravitino.integration.test.util.ITUtils;
 import org.apache.gravitino.storage.S3Properties;
-import org.apache.iceberg.exceptions.ForbiddenException;
-import org.apache.spark.SparkException;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 import org.junit.platform.commons.util.StringUtils;
 
 @SuppressWarnings("FormatStringAnnotation")
 @EnabledIfEnvironmentVariable(named = "GRAVITINO_TEST_CLOUD_IT", matches = "true")
-public class IcebergRESTS3TokenAuthorizationIT extends IcebergAuthorizationIT {
+public class IcebergRESTS3TokenAuthorizationIT extends IcebergRESTCloudTokenAuthorizationBaseIT {
 
-  private static final String SCHEMA_NAME = "schema";
   private String s3Warehouse;
   private String accessKey;
   private String secretKey;
@@ -78,30 +59,7 @@ public class IcebergRESTS3TokenAuthorizationIT extends IcebergAuthorizationIT {
 
     catalogClientWithAllPrivilege.asSchemas().createSchema(SCHEMA_NAME, "test", new HashMap<>());
 
-    if (ITUtils.isEmbedded()) {
-      return;
-    }
-    try {
-      downloadIcebergAwsBundleJar();
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    copyS3BundleJar();
-  }
-
-  @BeforeEach
-  void revokePrivilege() {
-    revokeUserRoles();
-    resetMetalakeAndCatalogOwner();
-    MetadataObject schemaObject =
-        MetadataObjects.of(
-            Arrays.asList(GRAVITINO_CATALOG_NAME, SCHEMA_NAME), MetadataObject.Type.SCHEMA);
-    metalakeClientWithAllPrivilege.setOwner(schemaObject, SUPER_USER, Owner.Type.USER);
-    clearTable();
-    // Grant user the privilege to use the catalog and schema
-    grantUseSchemaRole(SCHEMA_NAME);
-    sql("USE %s;", SPARK_CATALOG_NAME);
-    sql("USE %s;", SCHEMA_NAME);
+    setupCloudBundles();
   }
 
   @Override
@@ -112,179 +70,12 @@ public class IcebergRESTS3TokenAuthorizationIT extends IcebergAuthorizationIT {
   }
 
   @Override
-  protected boolean supportsCredentialVending() {
-    return true;
+  protected String getCloudProviderName() {
+    return "s3";
   }
 
-  @Test
-  void testIcebergOwnerS3Token() {
-    String tableName = "test_owner_s3";
-    grantCreateTableRole(SCHEMA_NAME);
-    sql("CREATE TABLE %s(a int, b int) PARTITIONED BY (a)", tableName);
-    sql("INSERT INTO %s VALUES (1,1),(2,2)", tableName);
-    List<Object[]> rows = sql("SELECT * FROM %s", tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s", SCHEMA_NAME, tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s.partitions", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT *,_file FROM %s", tableName);
-    Assertions.assertEquals(2, rows.size());
-  }
-
-  @Test
-  void testIcebergSelectTableS3Token() {
-    String tableName = "test_select_s3";
-    createTable(SCHEMA_NAME, tableName);
-
-    // No privileges
-    Assertions.assertThrows(
-        ForbiddenException.class, () -> sql("INSERT INTO %s VALUES (1,1),(2,2)", tableName));
-    Assertions.assertThrows(
-        ForbiddenException.class,
-        () -> sql("SELECT * FROM %s.%s.%s", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName));
-
-    grantSelectTableRole(tableName);
-    Assertions.assertThrows(
-        SparkException.class, () -> sql("INSERT INTO %s VALUES (1,1),(2,2)", tableName));
-    List<Object[]> rows = sql("SELECT * FROM %s", tableName);
-    Assertions.assertEquals(0, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(0, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s", SCHEMA_NAME, tableName);
-    Assertions.assertEquals(0, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s.partitions", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(0, rows.size());
-
-    rows = sql("SELECT *,_file FROM %s", tableName);
-    Assertions.assertEquals(0, rows.size());
-  }
-
-  @Test
-  void testIcebergModifyTableS3Token() {
-    String tableName = "test_modify_s3";
-    createTable(SCHEMA_NAME, tableName);
-
-    // No privileges
-    Assertions.assertThrows(
-        ForbiddenException.class, () -> sql("INSERT INTO %s VALUES (1,1),(2,2)", tableName));
-    Assertions.assertThrows(
-        ForbiddenException.class,
-        () -> sql("SELECT * FROM %s.%s.%s", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName));
-
-    grantModifyTableRole(tableName);
-
-    sql("INSERT INTO %s VALUES (1,1),(2,2)", tableName);
-    List<Object[]> rows = sql("SELECT * FROM %s", tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s", SCHEMA_NAME, tableName);
-    Assertions.assertEquals(2, rows.size());
-
-    rows = sql("SELECT * FROM %s.%s.%s.partitions", SPARK_CATALOG_NAME, SCHEMA_NAME, tableName);
-    Assertions.assertEquals(1, rows.size());
-
-    rows = sql("SELECT *,_file FROM %s", tableName);
-    Assertions.assertEquals(2, rows.size());
-  }
-
-  private void grantUseSchemaRole(String schema) {
-    String roleName = "useSchema_" + UUID.randomUUID();
-    List<SecurableObject> securableObjects = new ArrayList<>();
-    SecurableObject catalogObject =
-        SecurableObjects.ofCatalog(
-            GRAVITINO_CATALOG_NAME, ImmutableList.of(Privileges.UseCatalog.allow()));
-    securableObjects.add(catalogObject);
-    SecurableObject schemaObject =
-        SecurableObjects.ofSchema(
-            catalogObject, schema, ImmutableList.of(Privileges.UseSchema.allow()));
-    securableObjects.add(schemaObject);
-    metalakeClientWithAllPrivilege.createRole(roleName, new HashMap<>(), securableObjects);
-
-    metalakeClientWithAllPrivilege.grantRolesToUser(ImmutableList.of(roleName), NORMAL_USER);
-  }
-
-  private String grantCreateTableRole(String schema) {
-    String roleName = "createTable_" + UUID.randomUUID();
-    List<SecurableObject> securableObjects = new ArrayList<>();
-    SecurableObject catalogObject =
-        SecurableObjects.ofCatalog(
-            GRAVITINO_CATALOG_NAME, ImmutableList.of(Privileges.UseCatalog.allow()));
-    securableObjects.add(catalogObject);
-    SecurableObject schemaObject =
-        SecurableObjects.ofSchema(
-            catalogObject, schema, ImmutableList.of(Privileges.CreateTable.allow()));
-    securableObjects.add(schemaObject);
-    metalakeClientWithAllPrivilege.createRole(roleName, new HashMap<>(), securableObjects);
-    metalakeClientWithAllPrivilege.grantRolesToUser(ImmutableList.of(roleName), NORMAL_USER);
-    return roleName;
-  }
-
-  private String grantSelectTableRole(String tableName) {
-    String roleName = "selectTable_" + UUID.randomUUID();
-    List<SecurableObject> securableObjects = new ArrayList<>();
-    SecurableObject catalogObject =
-        SecurableObjects.ofCatalog(
-            GRAVITINO_CATALOG_NAME, ImmutableList.of(Privileges.UseCatalog.allow()));
-    securableObjects.add(catalogObject);
-    SecurableObject schemaObject =
-        SecurableObjects.ofSchema(
-            catalogObject, SCHEMA_NAME, ImmutableList.of(Privileges.UseSchema.allow()));
-    SecurableObject tableObject =
-        SecurableObjects.ofTable(
-            schemaObject, tableName, ImmutableList.of(Privileges.SelectTable.allow()));
-    securableObjects.add(tableObject);
-    metalakeClientWithAllPrivilege.createRole(roleName, new HashMap<>(), securableObjects);
-    metalakeClientWithAllPrivilege.grantRolesToUser(ImmutableList.of(roleName), NORMAL_USER);
-    return roleName;
-  }
-
-  private String grantModifyTableRole(String tableName) {
-    String roleName = "modifyTable_" + UUID.randomUUID();
-    List<SecurableObject> securableObjects = new ArrayList<>();
-    SecurableObject catalogObject =
-        SecurableObjects.ofCatalog(
-            GRAVITINO_CATALOG_NAME, ImmutableList.of(Privileges.UseCatalog.allow()));
-    securableObjects.add(catalogObject);
-    SecurableObject schemaObject =
-        SecurableObjects.ofSchema(
-            catalogObject, SCHEMA_NAME, ImmutableList.of(Privileges.UseSchema.allow()));
-    SecurableObject tableObject =
-        SecurableObjects.ofTable(
-            schemaObject, tableName, ImmutableList.of(Privileges.ModifyTable.allow()));
-    securableObjects.add(tableObject);
-    metalakeClientWithAllPrivilege.createRole(roleName, new HashMap<>(), securableObjects);
-    metalakeClientWithAllPrivilege.grantRolesToUser(ImmutableList.of(roleName), NORMAL_USER);
-    return roleName;
-  }
-
-  private void clearTable() {
-    Arrays.stream(
-            catalogClientWithAllPrivilege.asTableCatalog().listTables(Namespace.of(SCHEMA_NAME)))
-        .forEach(
-            table -> {
-              catalogClientWithAllPrivilege
-                  .asTableCatalog()
-                  .dropTable(NameIdentifier.of(SCHEMA_NAME, table.name()));
-            });
-    NameIdentifier[] nameIdentifiers =
-        catalogClientWithAllPrivilege.asTableCatalog().listTables(Namespace.of(SCHEMA_NAME));
-    Assertions.assertEquals(0, nameIdentifiers.length);
-  }
-
-  private void downloadIcebergAwsBundleJar() throws IOException {
+  @Override
+  protected void downloadCloudBundleJar() throws IOException {
     String icebergBundleJarUri =
         String.format(
             "https://repo1.maven.org/maven2/org/apache/iceberg/"
@@ -295,7 +86,8 @@ public class IcebergRESTS3TokenAuthorizationIT extends IcebergAuthorizationIT {
     DownloaderUtils.downloadFile(icebergBundleJarUri, targetDir);
   }
 
-  private void copyS3BundleJar() {
+  @Override
+  protected void copyCloudBundleJar() {
     String gravitinoHome = System.getenv("GRAVITINO_HOME");
     String targetDir = String.format("%s/iceberg-rest-server/libs/", gravitinoHome);
     BaseIT.copyBundleJarsToDirectory("aws", targetDir);
